+++ scglobals/static/tx/src/app/helpers/currency.js
@@ -1,4 +1,4 @@
 module.exports = function(amount) {
-  if (amount) return amount.toFixed(2).replace(/\B(?=(\d{3})+(?!\d))/g, ",");
+  if ((amount) && Number(amount)) return amount.toFixed(2).replace(/\B(?=(\d{3})+(?!\d))/g, ",");
   else return "0.0";
 };