+++ scglobals/static/tx/src/app/views/dash.js
@@ -17,22 +17,39 @@ export default Backbone.View.extend({
   },
   showShipments: function(){
     $("#loading").hide()
-    $("#shipments").removeClass('hidden').hide().fadeIn(1000);
+    $("#shipments").removeClass('hidden').hide().fadeIn(500);
   },
   showSummary: function(){
     $("#loading-summary").fadeOut('fast');
-    $("#shipments-summary").removeClass('hidden').hide().fadeIn(1000);
+    $("#shipments-summary").removeClass('hidden').hide().fadeIn(500);
   },
-  render: function() {
+  render: function(options) {
+    options = options || {};
     this.$el.html(this.template());
-    var shipments_promise = this.shipmentsCollection.fetch().then((shipments)=>{
+    var shipments_promise = this.shipmentsCollection.fetch(options).then((shipments)=>{
       $("#shipments").html(this.shipmentsTemplate(shipments.doc_rows));
       return shipments.total_rows;
     });
-    var shipments_value_promise = this.shipmentsCollection.fetch({reduce: true},'shipments-by-value');
-    var shipments_transactions_promise = this.shipmentsCollection.fetch({reduce: true},'shipments-by-count');
+    var shipments_value_promise = this.shipmentsCollection.fetch(options,'shipments-by-value',true);
+    var shipments_transactions_promise = this.shipmentsCollection.fetch(options,'shipments-by-count',true);
     Promise.all([shipments_promise,shipments_value_promise,shipments_transactions_promise]).then((results)=>{
-      $("#shipments-summary").html(ShipmentsSummaryTemplate(results));
+      var shipmentSummary = {
+        date_string: "",
+        total_shipments: results[0],
+        total_value: results[1],
+        total_transactions: results[2],
+        skip: options.skip || 0,
+        limit: options.limit || 1000,
+        previousURLParams: this.shipmentsCollection.makePreviousUrlParams(options),
+        nextURLParams: this.shipmentsCollection.makeNextUrlParams(options)
+      };
+      $("#shipments-summary").html(ShipmentsSummaryTemplate(shipmentSummary));
+      if (Number(shipmentSummary.skip) === 0){
+        $("#previous").addClass('disabled');
+      }
+      if (Number(shipmentSummary.skip) + Number(shipmentSummary.limit) > shipmentSummary.total_shipments){
+        $("#next").addClass('disabled');
+      }
       this.showSummary();
       this.showShipments();
     });