+++ scglobals/static/tx/changelog.md
@@ -6,6 +6,8 @@ User
 * sees 1000 shipments as date, from, to, total items, total value | Sm | 2h
 * can click a shipment and see the shipment's transactions | already have the shipment loaded on list view, so this could be one less hit to the DB... | sm | 1h
 * sees the total number of shipments, their total value, and their total number of items | Sm | 2h
+* sees pagination | Sm | 45m
+* can paginate through shipments | M | 2.5h
 
 #### improvements